feature: make a people tab
